Fix doc and finalize some formats
