+++ vqe/example_hamiltonians/README.md
@@ -1,5 +1,5 @@
 # Hamiltonians
 
 * `H415.hamil`: linear H4 (4 spartial orbitals, 4 particles), R(H-H) = 1.5 Angstrom, FCI energy -1.9961503255188
-* `H618.hamil`: linear H6 (6 spartial orbitals, 6 particles), R(H-H) = 1.8 Angstrom, FCI energy -2.88756937643578
-
+* `H618.hamil`: linear H6 (6 spartial orbitals, 6 particles), R(H-H) = 1.8 Angstrom, FCI energy -2.88756937643578.  The FCI energy from the [published source](10.1021/acs.jctc.9b01083) is -2.88756937438597.
+* `H622.hamil`: linear H6 (6 spartial orbitals, 6 particles), R(H-H) = 2.2 Angstrom, FCI energy -2.82434165191321. The FCI energy from the [published source](10.1021/acs.jctc.9b01083) is -2.82434165111152.
