+++ vqe/include/utils.hpp
@@ -95,10 +95,10 @@ namespace NWQSim{
     OptimizerSettings(): rel_tol(-1), 
                          abs_tol(-1),
                          stop_val(-MAXFLOAT),
-                         max_evals(200), // MZ: changed from 50 to 200, to match the default in main.cpp. And 200 is still too small for H4 level with COBYLA
+                         max_evals(100), // MZ: changed from 50 to 100, to match the default in main.cpp.
                          max_time(-1),
-                         lbound(-PI),
-                         ubound(PI) {}
+                         lbound(-PI+0.00001), // MZ: avoid bad numerical issues
+                         ubound(PI-0.00001) {}
   };
